Terminal state fix (#8)
There's a problem with stochastic MuZero: if we don't observe the terminal state, we cannot run the chance outcome encoder on it, which means we don't have the chance outcome for the final reward prediction. It is currently solved in codebase by forcing the chance outcome for the last action to be zero. This solution is definitely not correct as the last action can lead to different outcomes. Imagine fighting a monster: if you attack it, you can either die and go to a terminal state, or win and proceed.

This PR attempts to work around the problem. The idea is as follows: when we reach the original terminal state, we don't return "done", but actually transition into that state. Any action from the original terminal state will return "done" and give the reward of zero, so the final transition is now deterministic.

While this should help in theory, no profit has been observed in the experiments, so this change is optional and disabled by default.

This change also introduces a new "Retreat" environment, where large reward can be obtained by defeating a monster, but when the fight doesn't seem to go well, it makes more sense to retreat and get a smaller reward. This was intended as a testbed for correctly modelling monster encounters.

This change also refactors the way we work with gym environments, moving all logic intro env wrappers (and a separate lib).